Changed boolean
Initially, a boolean was asked for the traject shape, so it could only handle False. If false, it would get the traject shape from the waterveiligheidsportaal WFS. However, we want to be able to give a path. So it should check for a string to an existing path, otherwise give a False

To do: if trajectshape is not false, check if it is a path to an existing file. And does it have to be a shapefile or is geopackage also possible?